# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_comitanigiacomo_kanso-sync-engine_internal_core_domain.HabitEntry:
    properties:
      completion_date:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      habit_id:
        type: string
      id:
        type: string
      notes:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      value:
        type: integer
      version:
        type: integer
    type: object
  internal_adapters_handler_http.createEntryRequest:
    properties:
      completion_date:
        type: string
      habit_id:
        type: string
      notes:
        type: string
      value:
        type: integer
    required:
    - completion_date
    - habit_id
    type: object
...
  internal_adapters_handler_http.updateEntryRequest:
    properties:
      notes:
        type: string
      value:
        type: integer
      version:
        type: integer
    required:
    - version
    type: object
...
host: localhost:8080
info:
  contact:
    name: Giacomo Comitani
    url: https://github.com/comitanigiacomo
  description: |-
    Backend for the Kanso Habit Tracker application.
    Features include offline-first sync, streak calculation, and JWT authentication.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Kanso Sync Engine API
  version: "1.0"
paths:
...
  /entries:
    get:
      description: Get history of entries for a specific habit ID within a date range
      parameters:
      - description: Habit ID
        in: query
        name: habit_id
        required: true
        type: string
      - description: 'Start Date (RFC3339) - Default: 30 days ago'
        in: query
        name: from
        type: string
      - description: 'End Date (RFC3339) - Default: Now'
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_comitanigiacomo_kanso-sync-engine_internal_core_domain.HabitEntry'
            type: array
        "400":
          description: Missing habit_id
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List entries for a habit
      tags:
      - Entries
    post:
      consumes:
      - application/json
      description: Record a completion or value for a specific habit on a specific
        date
      parameters:
      - description: Entry Data
        in: body
        name: entry
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_handler_http.createEntryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_comitanigiacomo_kanso-sync-engine_internal_core_domain.HabitEntry'
        "400":
          description: Invalid Input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Log a habit entry
      tags:
      - Entries
  /entries/{id}:
    delete:
      description: Permanently remove an entry
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Entry not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete an entry
      tags:
      - Entries
    put:
      consumes:
      - application/json
      description: Change the value or completion status. Requires current version
        for optimistic locking.
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Data
        in: body
        name: entry
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_handler_http.updateEntryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_comitanigiacomo_kanso-sync-engine_internal_core_domain.HabitEntry'
        "400":
          description: Invalid Input
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Entry not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Version Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an entry value
      tags:
      - Entries
  /entries/sync:
    get:
      description: Get entries created or modified since the last sync timestamp
      parameters:
      - description: Timestamp (RFC3339)
        in: query
        name: since
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'Returns {changes: [], timestamp: ...}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid Date Format
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Sync entries (Offline-First)
      tags:
      - Entries
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"